Update patch set 1
Patch Set 1:

> > Prefer to not include the oxm file here, or is it necessary for
 > > these test cases?  It's just one more place where it has to be
 > > synced.
 > 
 > we need the oxm file for successful execution of the test cases.
 > Regarding the place of oxm file, kindly let me know if moving it
 > under src/test/resources would be better, So that i can go ahead
 > and do that. Thanks

Hi Aditya,
I think what Jimmy is trying to do is to avoid a copy of oxm file in the code base and get it from aai-common in
https://gerrit.onap.org/r/gitweb?p=aai/aai-common.git;a=tree;f=aai-schema/src/main/resources/oxm;h=ebdf665a53dd5140293d8c347fa6dadaaaa7c93b;hb=refs/heads/master

 > > Prefer to not include the oxm file here, or is it necessary for
 > > these test cases?  It's just one more place where it has to be
 > > synced.
 > 
 > we need the oxm file for successful execution of the test cases.
 > Regarding the place of oxm file, kindly let me know if moving it
 > under src/test/resources would be better, So that i can go ahead
 > and do that. Thanks
Hi Aditya,
I think what Jimmy is saying is that we should not have a oxm file in the repo and it does not matter where it is located. It should be retrieved form
https://gerrit.onap.org/r/gitweb?p=aai/aai-common.git;a=tree;f=aai-schema/src/main/resources/oxm;h=ebdf665a53dd5140293d8c347fa6dadaaaa7c93b;hb=refs/heads/master

Either we do that or remove the test cases that depend on the file. 
regards
Arul

Patch-set: 1