Update patch set 2
Patch Set 2:

Build Failed 

https://jenkins.onap.org/job/aai-sparky-be-amsterdam-aai-docker-java-daily/36/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch ( https://jenkins.onap.org/job/aai-sparky-be-amsterdam-aai-docker-java-daily/36/ )

Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/aai-sparky-be-amsterdam-aai-docker-java-daily/36

https://jenkins.onap.org/job/aai-sparky-be-amsterdam-release-version-java-daily/36/ : FAILURE

Unable to build due to a non-resolvable parent POM. Search logs for "[FATAL] Non-resolvable parent POM" to find the logs.

The fix for this is usually to make sure your poms are pointing to parents that exist. Double check:

 * groupId is correct and no typos
 * artifactId is correct and no typos
 * version is correct for branch ( https://jenkins.onap.org/job/aai-sparky-be-amsterdam-release-version-java-daily/36/ )

Logs: https://logs.onap.org/production/vex-yul-ecomp-jenkins-1/aai-sparky-be-amsterdam-release-version-java-daily/36

Patch-set: 2