Update patch set 1
Patch Set 1: Code-Review-1

I agree with reading the properties files instead of using a constant value which is surely a bug, but a minor one. And I think we can address it in the next release as Beijing code is frozen at this point.
For the second part where we want to add a new property, i dont agree with it because adding 2 properties which does the same thing does not make sense, and maybe the solution is to change the existing days to minutes. We can discuss this.

Patch-set: 1
Reviewer: Gerrit User 1041 <1041@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1