Update patch set 1
Patch Set 1:

> @Arul, how is this working in the other envs?  Using different
 > settings?

I do see this issue in the logs in windriver. What @Ashok found is right, there is a chance that the headers can be duplicated with this code. And To fix it instead of commenting out the whole code, I would check if the header key is present in the map first and if the key is absent I would add the header. This will be a better fix rather than comment the code out, incase if other classes are using it. 
eg
Object headerKey = headers.get(HEADER_AUTHORIZATION);
if (headerKey != null) {
    headers.putIfAbsent(HEADER_AUTHORIZATION, new ArrayList<String>());
      headers.get(HEADER_AUTHORIZATION).add(getBasicAuthenticationCredentials());
}


Also there are unwanted imports which the commit, which needs to be cleaned up.

Thanks @Ashok

Patch-set: 1